+++ bridge/server/utils/print-utils.spec.ts
@@ -1,8 +1,9 @@
 import { AxiosError, Method } from 'axios';
 // eslint-disable-next-line import/no-extraneous-dependencies
 import { jest } from '@jest/globals';
-import { printError } from './print-utils';
+import { filterHeaders, printError } from './print-utils';
 import { LogDestination, logger } from './logger';
+import { IncomingHttpHeaders } from 'http';
 
 describe('Test print-utils', () => {
   beforeAll(() => {
@@ -55,5 +56,49 @@ describe('Test print-utils', () => {
     printError(error);
     expect(consoleSpy).toBeCalledTimes(1);
     expect(consoleSpy).toHaveBeenCalledWith('API', `Error: ${msg}`);
+  });
+
+  it('should not filter anything', () => {
+    const headers: IncomingHttpHeaders = {
+      host: 'testytest',
+      path: 'testytest',
+      'content-type': 'testytest',
+      'www-authenticate': 'testytest',
+      myheader: 'testytest',
+    };
+    expect(filterHeaders(headers)).toStrictEqual(headers);
+  });
+
+  it('should filter sec headers', () => {
+    const headers: IncomingHttpHeaders = {
+      host: 'testytest',
+      path: 'testytest',
+      'content-type': 'testytest',
+      'www-authenticate': 'testytest',
+      myheader: 'testytest',
+    };
+    const inputHeaders = {
+      ...headers,
+      'sec-fetch-dest': 'testytest',
+      'sec-fetch-mode': 'testytest',
+      'sec-ch-ua': 'testytest',
+    };
+    expect(filterHeaders(inputHeaders)).toStrictEqual(headers);
+  });
+
+  it('should well known headers', () => {
+    const headers: IncomingHttpHeaders = {
+      host: 'testytest',
+      path: 'testytest',
+      'content-type': 'testytest',
+      'www-authenticate': 'testytest',
+      myheader: 'testytest',
+    };
+    const inputHeaders = {
+      ...headers,
+      'user-agent': 'testytest',
+      cookies: 'testytest',
+    };
+    expect(filterHeaders(inputHeaders)).toStrictEqual(headers);
   });
 });
